update embed code
