+++ src/utils/events.js
@@ -23,7 +23,8 @@ const eventList = {
     CHART_BLOCKS: 'CHART_BLOCKS',
     CHART_PUBLISH_DATA: 'CHART_PUBLISH_DATA',
     PLUGINS_LOADED: 'PLUGINS_LOADED',
-    EXTERNAL_DATA_URL: 'EXTERNAL_DATA_URL'
+    EXTERNAL_DATA_URL: 'EXTERNAL_DATA_URL',
+    GET_NEXT_PUBLIC_URL: 'GET_NEXT_PUBLIC_URL'
 };
 
 module.exports = { ApiEventEmitter, eventList };
